apps/survey/requirements.txt
alembic==1.13.3
async-timeout==4.0.3
asyncpg==0.29.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-tools==1.66.2
Mako==1.3.5
MarkupSafe==3.0.1
protobuf==5.28.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
